fixed hight; made test src folder tidyer

+++ src/ciphers/hight/hight_expansion.h
@@ -6,7 +6,7 @@
 #include "hight_state.h"
 
 
-void hight_expandkey(uint8_t *user_key, uint32_t *key_enc, uint32_t *key_dec, size_t keylen);
+void hight_expandkey(uint8_t *user_key, uint8_t *key_enc, uint8_t *key_dec, size_t keylen);
 
 
 #endif
